🏗️ Fix archi
