+++ lib/domain/repositories/website_versions.repository.dart
@@ -1,0 +1,5 @@
+import 'package:aeweb/model/website_version.dart';
+
+abstract class WebsiteVersionsRepository {
+  Future<List<WebsiteVersion>> getWebsiteVersions(String genesisAddress);
+}
