+++ lib/domain/repositories/websites.repository.dart
@@ -1,0 +1,5 @@
+import 'package:aeweb/model/website.dart';
+
+abstract class WebsitesRepository {
+  Future<List<Website>> getWebsites();
+}
