+++ lib/ui/views/unpublish_website/layouts/components/unpublish_website_in_progress_popup.dart
@@ -1,5 +1,5 @@
 /// SPDX-License-Identifier: AGPL-3.0-or-later
-import 'package:aeweb/application/websites.dart';
+import 'package:aeweb/application/website_versions.dart';
 import 'package:aeweb/domain/usecases/website/unpublish_website.usecase.dart';
 import 'package:aeweb/ui/themes/aeweb_theme_base.dart';
 import 'package:aeweb/ui/views/unpublish_website/bloc/provider.dart';
@@ -260,7 +260,7 @@ class UnpublishWebsiteInProgressPopup {
                                   : '',
                               warningCloseFunction: () async {
                                 ref.invalidate(
-                                  WebsitesProviders.fetchWebsiteVersions,
+                                  WebsiteVersionProviders.fetchWebsiteVersions,
                                 );
 
                                 ref
